moved public, component files
